v2/flutter-package-registry/data/functions/pub_scraper/requirements.txt
firebase_functions~=0.1.0
functions-framework
google-cloud-firestore
beautifulsoup4
requests
parsedatetime
python-dateutil